Add fully functioning movie fragment
There was a _lot_ of experimentation to get to here. Unfortunately, I checked
in an API key, and didn't think it was a big deal since I could just regenerate
a new one after I pushed this up. Surprise! The omdb API doesn't allow you to
generate a new key. Soo I either had to meticulously scrub the git history or
just nuke it. I went the Ghandi route.